+++ frontends/web/src/components/balance/balance.test.tsx
@@ -48,11 +48,13 @@ describe('components/balance/balance', () => {
       hasIncoming: true,
       available: {
         amount: '0.005',
-        unit: 'BTC'
+        unit: 'BTC',
+        estimated: false,
       },
       incoming: {
         amount: '0.003',
         unit: 'BTC',
+        estimated: false,
         conversions: {
           BTC: '0.003',
           AUD: '512',
@@ -98,11 +100,13 @@ describe('components/balance/balance', () => {
       hasIncoming: true,
       available: {
         amount: '0.005',
-        unit: 'BTC'
+        unit: 'BTC',
+        estimated: false,
       },
       incoming: {
         amount: '0.003',
         unit: 'BTC',
+        estimated: false,
         conversions: {
           BTC: '0.003',
           AUD: '512',